Eigenvector Clustering

Eigenvector Clustering

Once the eigenvectors are obtained, we have a continuous solution for a discrete problem. In order to obtain an assigment for every pattern, it is needed to discretize the eigenvectors. Obtaining this discrete solution from eigenvectors often requires solving another clustering problem, albeit in a lower-dimensional space. That is, eigenvectors are treated as geometrical coordinates of a point set.

This library provides two methods two obtain the discrete solution:

Kmeans by means of Clustering.jl

The one proposed in Multiclass spectral clustering

Examples

Eigenvector clusterization examples

Reference Index

Members Documentation

SpectralClustering.KMeansClusterizer — Type.

struct KMeansClusterizer <: EigenvectorClusterizer
    k::Integer
    init::Symbol
end

source

SpectralClustering.YuEigenvectorRotation — Type.

Multiclass Spectral Clustering

source

SpectralClustering.clusterize — Method.

function clusterize{T<:EigenvectorEmbedder, C<:EigenvectorClusterizer}(cfg::T, clus::C, X)

Given a set of patterns X generates an eigenvector space according to T<:EigenvectorEmbeddder and then clusterize the eigenvectors using the algorithm defined by C<:EigenvectorClusterize.

source

Bibliography

SS003

X Yu Stella and Jianbo Shi. Multiclass spectral clustering. In null. IEEE, 2003.