Graph Matrices

Graph matrices

The Graphs.Matrices module provides a collecion of functions that computes differents matrices from a graph. Given a Graph $G=(V,E)$ where $(v_i,v_j,w_{ij}) \in E, v_i,v_j \in V, w_{ij} \in R$ this module allows to obtain:

The simmiliarty matrix $W$. Where $W_{ij} = w_{ij}$

The laplacian matrix $L=(D-W)$. Where $D = W \mathbf{1}$

The normalized laplacian $L_N = D^{-\frac{1}{2}} (D-W) D^{-\frac{1}{2}}$

NSJ Laplacian $L_{NSJ} = D^{-1}W$

Detailed Description

Laplacian

Normalized Laplacian

NSJ Laplacian

Reference

Index

Content

SpectralClustering.degrees — Method.

degrees(g::Graph)

Returns a vector with the degrees of every vertex

source

SpectralClustering.incidence — Method.

incidence(g::Graph)

Returns the incidence matrix of the graph ´´´G´´´

source

SpectralClustering.laplacian — Method.

laplacian(W::SparseMatrixCSC, d::Vector)

Returns the laplacian matrix ´´´L´´´ of the graph ´´´g´´´

source

SpectralClustering.laplacian — Method.

laplacian(g::Graph)

Returns the laplacian matrix ´´´L´´´ of the graph ´´´g´´´

\[L=D-W\]

source

SpectralClustering.ng_laplacian — Method.

ng_laplacian(g::Graph)

Returns the laplacian matrix of the graph ´´´G´´´ according to ""On Spectral Clustering: Analysis and an algorithm"" of Andrew Y. Ng, Michael I. Jordan and Yair Weiss.

\[L = D^{-\frac{1}{2}} W D^{-\frac{1}{2}}\]

source

SpectralClustering.normalized_laplacian — Method.

normalized_laplacian(W::SparseMatrixCSC,d)

source

SpectralClustering.normalized_laplacian — Method.

 normalized_laplacian(W::SparseMatrixCSC)

source

SpectralClustering.normalized_laplacian — Method.

normalized_laplacian(g::Graph)

If W it is the weight matrix $W$ of ´´´g´´´, $D=W mathbf{1}$. The normalized laplacian $L = D^{-\frac{1}{2}} (D-W) D^{-\frac{1}{2}}$. The function returns the normalized laplacian of the graph ´´´g´´´ and $D^{\frac{1}{2}}$.

source

SpectralClustering.weight_matrix — Method.

weight_matrix(g::Graph)

Given a graph ´´´g´´´ and two array of vertices (set1, set2) indexes returns the weight matrix form the vertices in set1 to the vertices in set2.

source

SpectralClustering.weight_matrix — Method.

weight_matrix(g::Graph)

Given a graph ´g´ returns the weight matrix and a vector with the vertices degrees.

source

SpectralClustering.weight_matrix — Method.

julia weight_matrix{T<:AbstractFloat}(n_type::T,g::Graph)`

Given a graph ´g´ returns the weight matrix of `n_type and a vector with the vertices degrees.

source