Data Access

Data Access

In order to establish how the data is going to be accessed, the module DataAccess provides an unified interface to access to the data for the underlaying algorithms. Every DataAccessor must implement this two methods:

get_element(d::T, X, i::Integer). This function must return the i-th pattern of X.

number_of_patterns(d::T,X). This function must return the numer of patterns of X

Reference Index

Members Documentation

SpectralClustering.assign! — Method.

function assign!(vec::T, val::C) where T<:AbstractArray where C<:Colorant

This function assigns the components of the color component val to a vector v

source

SpectralClustering.get_element! — Method.

function get_element!(o::Matrix,  img::Matrix{C}, i::Vector{Integer}) where C<:Colorant

source

SpectralClustering.get_element! — Method.

get_element!{T<:AbstractArray}(vec::T,  img::Matrix{Gray}, i::Integer)

Return throughvec the intensity image element [x,y, i], where $x,y$ are the spatial position of the pixel and the value i of the pixel $(x,y)$.

source

SpectralClustering.get_element — Method.

function get_element( img::Matrix{RGB}, i::Vector)

source

SpectralClustering.get_element — Method.

function get_element(data::Array{T, 3}, i::Vector)  where T<:Number

source

SpectralClustering.number_of_patterns — Method.

number_of_patterns{T<:Any}(X::Matrix{T,3})

Return the number of pixels in the image

source

SpectralClustering.spatial_position — Method.

 spatial_position(X::Matrix, i::Int)

Returns the sub indexes from the linear index i

source