Co-Regularized

MultiView Embedding

When the dataset has more than one representation, each of them is named view. In the context of spectral clustering, co-regularization techniques attempt to encourage the similarity of the examples in the new representation generated from the eigenvectors of each view.

Examples

MultiView Embedding examples

Reference Index

Members Documentation

SpectralClustering.CoRegularizedMultiView — Type.

Co-regularized Multi-view Spectral Clustering

Abhishek Kumar, Piyush Rai, Hal Daumé

source

SpectralClustering.View — Type.

A view

struct View
    embedder::NgLaplacian
    lambda::Float64
end

Members

- embedder::EigenvectorEmbedder
- lambda::Float64

source

SpectralClustering.embedding — Method.

embedding(cfg::CoRegularizedMultiView, X::Vector)

Arguments

- `cfg::CoRegularizedMultiView`
- `X::Vector{Graph}`

An example that shows how to use this methods is provied in the Usage section of the manual

source

SpectralClustering.LargeScaleMultiView — Type.

type LargeScaleMultiView

Large-Scale Multi-View Spectral Clustering via Bipartite Graph. In AAAI (pp. 2750-2756).

Li, Y., Nie, F., Huang, H., & Huang, J. (2015, January).

Matlab implementation

Members

k::Integer. Number of clusters.

n_salient_points::Integer. Number of salient points.

k_nn::Integer. k nearest neighbors.

'gamma::Float64`.

source